Send Welcome Mailer - AR Callback
This changes adds in the functionality to send the welcome mailer upon a
study participant's enrollment in a study. This is done by adding an
`after_commit` callback on the participant model, so that every time
that a participant is created, through whatever means, the AR callback
should fire and result in sending the email.