Send Welcome Mailer - Orchestrating Class
This provides an alternative for sending a welcome mailer upon a study
participant's enrollment in a study. It does so by introducing a new
ruby class that's singularly responsible for conducting the enrollment
process, which involves creating a study participant and then sending
them the welcome mailer.

This allows for the business language of the event (enrollment) to be
used, and also decouples the sending of the welcome mailer from the
model's AR lifecycle directly. It can still be enforced to be sent by
passing through the enrollment class; however, one can also make a
participant without sending the mailer.