Warn user about editing personal information - unit test
This also warns a user that editing their personal information may
affect their eligibility. However, given the norm in the code base that
flash messages are not verified via system tests due to the expense of
running a system for such a trivial feature, that spec has been removed.

Given the importance of the specifics of that message, in order to
comply with FDA regulations, the `PersonalInformation` class will now
respond with that message, which the controller will use to set the
flash message. That method on the `PersonalInformation` class is tested
instead, so there can be coverage that the message is as expected.