Warn user about editing personal information - append system test
This validates the flash message warning a user that editing a
participants personal information may affect their eligibility with a
system test. However, rather than a separate test solely focused on it,
the verification is added to an existing spec that is focused on the
overall flow for editing personal information.

Even if the code base's convention is to not test flash messages, the
importance and the impact that not having this flash message has from a
regulatory compliance perspective may outweigh the convention in the
application's followed practices.