Extract named attribute for when collection event took place
This removes the reliance on the rails timestamp for when a data
collection event took place, and instead extracts a separate attribute
to be used in its place. This provides for a named value that's
consistent with business terminology.

In addition, it separates out the uses or meaning of the `created_at`
value. Previously, should we need to create data collection events
before they took place, or if we ever needed to receive a data import
of legacy data, we would need to decide which is more important:
knowing when the event took place, or when the event was actually
created in the database.