Only Allow Principal Investigators Access to Create Protocol - Pundit
This introduces an external dependency, pundit, to manage the
authorization of a user being able to create a new study protocol. This
extracts that logic out of the controller, and calls out to it in the
controller, along with some custom error-handling to maintain the
existing functionality about the flash message and routing.

This provides a pattern that is extensible, may be used elsewhere, and
may be utilized by other objects, but comes at the cost of utilizing
a third-party dependency, and indirection in the controller.